Add the ability to pass in values to append to end of branch name
I used to have the flexibility to append values to the end of a gernerated
branch name. Recently the interface was improved so there is now only a single
command, but I still like the ability to append those values. For example, if
you have multiple branches for a single story on Pivotal, it is nice to do something like:
```
$ pv-git-branch 123456 backend
aaron+fry/features/allow-pv-git-branch-to-pass-values-123456-backend
$ pv-git-branch 123456 frontend
aaron+fry/features/allow-pv-git-branch-to-pass-values-123456-frontend
```
... or if you have multiple versions of that branch for a single story:
```
$ pv-git-branch 123456 v1 spike
aaron+fry/features/allow-pv-git-branch-to-pass-values-123456-v1-spike
$ pv-git-branch 123456 v2 final
aaron+fry/features/allow-pv-git-branch-to-pass-values-123456-v2-final
```
This is backward compatible, so you can still run the command the old way with
now appending values:
```
$ pv-git-branch 123456
aaron+fry/features/allow-pv-git-branch-to-pass-values-123456
```

[#164945069]